- Fix a number of typecast issues. - Loosen timing assertions on SM unit test.

+++ ssfjson.h
@@ -121,9 +121,9 @@ bool SSFJsonPrintBase64(SSFCStrOut_t jstr, size_t size, size_t start, size_t *en
 bool SSFJsonPrintDouble(SSFCStrOut_t js, size_t size, size_t start, size_t *end, double in,
                         SSFJsonFltFmt_t fmt, bool *comma);
 #endif /* SSF_JSON_CONFIG_ENABLE_FLOAT_GEN */
-bool SSFJsonPrintInt(SSFCStrOut_t js, size_t size, size_t start, size_t *end, int32_t in,
+bool SSFJsonPrintInt(SSFCStrOut_t js, size_t size, size_t start, size_t *end, long in,
                      bool *comma);
-bool SSFJsonPrintUInt(SSFCStrOut_t js, size_t size, size_t start, size_t *end, uint32_t in,
+bool SSFJsonPrintUInt(SSFCStrOut_t js, size_t size, size_t start, size_t *end, unsigned long in,
                       bool *comma);
 #define SSFJsonPrintObject(js, size, start, end, fn, in, comma) \
         SSFJsonPrint(js, size, start, end, fn, in, "{}", comma)
